requirements.txt
numpy
networkx
matplotlib
scipy
psycopg2
sqlalchemy
pandas
black
python-dotenv